backend/requirements.txt
# API
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19
python-dotenv==1.1.0

# Azure Services
azure-ai-formrecognizer==3.3.3
azure-ai-textanalytics==5.3.0
azure-search-documents==11.6.0
azure-identity==1.21.0
openai==1.82.0

# Data Processing
pandas==2.2.3
openpyxl==3.1.5
python-docx==1.1.2
Pillow==11.2.1

# Document Chunking
docling==2.31.0
docling-core==2.31.0

# Neo4j
neo4j==5.28.1

# Utilities
httpx==0.28.1
tiktoken==0.9.0
numpy==2.2.6
